Struct awesome_bot::SendAudio

pub struct SendAudio {
    // some fields omitted
}

Transparent struct builded by SendBuilder to send Audio messages.

Methods

impl SendAudio

fn duration(&mut self, v: Integer) -> &mut SendAudio

Set the duration of the track

fn performer(&mut self, v: String) -> &mut SendAudio

Set the performer of the track

fn title(&mut self, v: String) -> &mut SendAudio

Set the title of the track

fn reply_id(&mut self, v: Integer) -> &mut SendAudio

Set a message ID to reply with this message.

fn markup(&mut self, v: ReplyMarkup) -> &mut SendAudio

Set a ReplyMarkup to send, but instead of this, use the keyboard, hide or force methods

impl SendAudio

Add keyboard methods to the struct, only one of this will be sended, and will be the last one used.

fn keyboard(&mut self, r: ReplyKeyboardMarkup) -> &mut SendAudio

Add a keyboard to the reply

fn hide(&mut self, h: bool) -> &mut SendAudio

Hide the keyboard

fn force(&mut self, f: bool) -> &mut SendAudio

Force the reply to this message