Struct awesome_bot::SendBuilder

pub struct SendBuilder { /* fields omitted */ }

SendBuilder it's a builder struct that allows you to construct answers in a composable way, you will use it transparently with the send and answer methods of AwesomeBot

Methods

impl SendBuilder

pub fn new(id: Integer, bot: Api) -> SendBuilder

Create a new SendBuilder, don't use it, use the send and answer methods of AwesomeBot :)

pub fn text(self, t: &str) -> SendText

Start a text constructor to send.

pub fn photo(self, t: &str) -> SendPhoto

Start a photo constructor to send.

pub fn audio(self, t: &str) -> SendAudio

Start an audio constructor to send.

pub fn voice(self, t: &str) -> SendVoice

Start a voice constructor to send.

pub fn document(self, t: &str) -> SendDocument

Start a document constructor to send.

pub fn sticker(self, t: &str) -> SendSticker

Start a sticker constructor to send.

pub fn video(self, t: &str) -> SendVideo

Start a video constructor to send.

pub fn forward(self, to: Integer, msg: Integer) -> SendForward

Start a forward constructor to send.

pub fn action(self, action: ChatAction) -> SendAction

Start an action constructor to send.

pub fn location(self, latitude: Float, longitude: Float) -> SendLocation

Start a location constructor to send.