Trait awesome_bot::Ender

pub trait Ender<T: Decodable> {
    fn end(&mut self) -> Result<T>;
}

Help trait that allows to type that at least the end method is implemented for the SendBuilder structs

Required Methods

fn end(&mut self) -> Result<T>