Struct awesome_bot::SendBuilder

pub struct SendBuilder {
    // some fields omitted
}

SendBuilder it's a builder struct that allows you to construct answers in a beauty way, you will use it transparently with the send and answer methods of AwesomeBot

Methods

impl SendBuilder

fn new(id: Integer, bot: Api) -> SendBuilder

Create a new SendBuilder, you shouldn't use it, use the send and answer methods of AwesomeBot :)

fn text(self, t: &str) -> SendText

Start a text constructor to send.

fn photo(self, t: &str) -> SendPhoto

Start a photo constructor to send.

fn audio(self, t: &str) -> SendAudio

Start an audio constructor to send.

fn voice(self, t: &str) -> SendVoice

Start a voice constructor to send.

fn document(self, t: &str) -> SendDocument

Start a document constructor to send.

fn sticker(self, t: &str) -> SendSticker

Start a sticker constructor to send.

fn video(self, t: &str) -> SendVideo

Start a video constructor to send.

fn forward(self, to: Integer, msg: Integer) -> SendForward

Start a forward constructor to send.

fn action(self, action: ChatAction) -> SendAction

Start an action constructor to send.

fn location(self, latitude: Float, longitude: Float) -> SendLocation

Start a location constructor to send.